requirements.txt
asgiref==3.8.1
Django==5.1.3
django-jazzmin==3.0.1
gunicorn==23.0.0
packaging==24.2
python-dotenv==1.0.1
sqlparse==0.5.2
django-import-export